lbs::acceleration::LBSSolver is in 1 .h/.cc file
