+++ modules/LinearBoltzmannSolvers/A_LBSSolver/lbs_solver.h
@@ -51,7 +51,452 @@ public:
   typedef std::shared_ptr<AGSLinearSolver<Mat, Vec, KSP>> AGSLinSolverPtr;
   typedef std::shared_ptr<chi_math::LinearSolver<Mat, Vec, KSP>> LinSolvePtr;
 
+public:
+  /**
+   * Returns the input parameters for this object.
+   */
+  static chi::InputParameters GetInputParameters();
+  explicit LBSSolver(const std::string& text_name);
+  /**
+   * Input parameters based construction.
+   */
+  explicit LBSSolver(const chi::InputParameters& params);
+
+  LBSSolver(const LBSSolver&) = delete;
+  LBSSolver& operator=(const LBSSolver&) = delete;
+
+  virtual ~LBSSolver() = default;
+
+  /**
+   * Returns the source event tag used for logging the time it takes to set source moments.
+   */
+  size_t GetSourceEventTag() const;
+
+  /**
+   * Returns the time at which the last restart was written.
+   */
+  double LastRestartWrite() const;
+
+  /**
+   * Returns a reference to the time at which the last restart was written.
+   */
+  double& LastRestartWrite();
+
+  /**
+   * Returns a reference to the solver options.
+   */
+  lbs::Options& Options();
+
+  /**
+   * Returns a constant reference to the solver options.
+   */
+  const lbs::Options& Options() const;
+
+  static chi::InputParameters OptionsBlock();
+  static chi::InputParameters BoundaryOptionsBlock();
+  void SetOptions(const chi::InputParameters& params);
+  void SetBoundaryOptions(const chi::InputParameters& params);
+
+  /**
+   * Returns the number of moments for the solver. This will only be non-zero after initialization.
+   */
+  size_t NumMoments() const;
+
+  /**
+   * Returns the number of groups for the solver. This will only be non-zero after initialization.
+   */
+  size_t NumGroups() const;
+
+  /**
+   * Returns the number of precursors for the solver. This will only be non-zero after
+   * initialization.
+   */
+  size_t NumPrecursors() const;
+
+  /**
+   * Returns the maximum number of precursors, for a material, as encountered accross all the
+   * materials. This will only be non-zero after initialization.
+   */
+  size_t GetMaxPrecursorsPerMaterial() const;
+
+  /**
+   * Adds a group to the list of groups. If group id < 0, the id will be logically derived from the
+   * list size. If >= 0 the id will be set to the id specified.
+   */
+  void AddGroup(int id);
+  const std::vector<LBSGroup>& Groups() const;
+
+  /**
+   * Adds a groupset to the list of groupsets. The groupset id will be logically
+   * derived from the list size.
+   */
+  void AddGroupset();
+  std::vector<LBSGroupset>& Groupsets();
+  const std::vector<LBSGroupset>& Groupsets() const;
+
+  /**
+   * Adds a point source to the solver's point source list.
+   */
+  void AddPointSource(PointSource psrc);
+
+  /**
+   * Clears all the point sources from the solver's point source list.
+   */
+  void ClearPointSources();
+
+  /**
+   * Const accessor to the list of point sources.
+   */
+  const std::vector<PointSource>& PointSources() const;
+
+  /**
+   * Returns a reference to the map of material ids to XSs.
+   */
+  const std::map<int, XSPtr>& GetMatID2XSMap() const;
+
+  /**
+   * Returns a reference to the map of material ids to Isotropic Srcs.
+   */
+  const std::map<int, IsotropicSrcPtr>& GetMatID2IsoSrcMap() const;
+
+  /**
+   * Obtains a reference to the spatial discretization.
+   */
+  const chi_math::SpatialDiscretization& SpatialDiscretization() const;
+
+  /**
+   * Returns read-only access to the unit cell matrices.
+   */
+  const std::vector<UnitCellMatrices>& GetUnitCellMatrices() const;
+
+  /**
+   * Obtains a reference to the grid.
+   */
+  const chi_mesh::MeshContinuum& Grid() const;
+
+  /**
+   * Returns a reference to the list of local cell transport views.
+   */
+  const std::vector<lbs::CellLBSView>& GetCellTransportViews() const;
+
+  /**
+   * Read/Write access to the boundary preferences.
+   */
+  std::map<uint64_t, BoundaryPreference>& BoundaryPreferences();
+
+  /**
+   * Obtains a reference to the unknown manager for flux-moments.
+   */
+  const chi_math::UnknownManager& UnknownManager() const;
+
+  /**
+   * Returns the local node count for the flux-moments data structures.
+   */
+  size_t LocalNodeCount() const;
+
+  /**
+   * Returns the global node count for the flux-moments data structures.
+   */
+  size_t GlobalNodeCount() const;
+
+  /**
+   * Read/write access to source moments vector.
+   */
+  std::vector<double>& QMomentsLocal();
+
+  /**
+   * Read access to source moments vector.
+   */
+  const std::vector<double>& QMomentsLocal() const;
+
+  /**
+   * Read/write access to exterior src moments vector.
+   */
+  std::vector<double>& ExtSrcMomentsLocal();
+
+  /**
+   * Read access to exterior src moments vector.
+   */
+  const std::vector<double>& ExtSrcMomentsLocal() const;
+
+  /**
+   * Read/write access to last updated flux vector.
+   */
+  std::vector<double>& PhiOldLocal();
+
+  /**
+   * Read access to last updated flux vector.
+   */
+  const std::vector<double>& PhiOldLocal() const;
+
+  /**
+   * Read/write access to newest updated flux vector.
+   */
+  std::vector<double>& PhiNewLocal();
+
+  /**
+   * Read access to newest updated flux vector.
+   */
+  const std::vector<double>& PhiNewLocal() const;
+
+  /**
+   * Read/write access to newest updated precursors vector.
+   */
+  std::vector<double>& PrecursorsNewLocal();
+
+  /**
+   * Read access to newest updated precursors vector.
+   */
+  const std::vector<double>& PrecursorsNewLocal() const;
+
+  /**
+   * Read/write access to newest updated angular flux vector.
+   */
+  std::vector<VecDbl>& PsiNewLocal();
+
+  /**
+   * Read access to newest updated angular flux vector.
+   */
+  const std::vector<VecDbl>& PsiNewLocal() const;
+
+  /**
+   * Returns the sweep boundaries as a read only reference
+   */
+  const std::map<uint64_t, std::shared_ptr<SweepBndry>>& SweepBoundaries() const;
+
+  SetSourceFunction GetActiveSetSourceFunction() const;
+
+  AGSLinSolverPtr GetPrimaryAGSSolver();
+
+  std::vector<LinSolvePtr>& GetWGSSolvers();
+
+  WGSContext<Mat, Vec, KSP>& GetWGSContext(int groupset_id);
+
+  /**
+   * Gets the local and global number of iterative unknowns. This normally is only the flux moments,
+   * however, the sweep based solvers might include delayed angular fluxes in this number.
+   */
+  virtual std::pair<size_t, size_t> GetNumPhiIterativeUnknowns();
+
+  /**
+   * Gets the local handle of a flux-moment based field function.
+   */
+  size_t MapPhiFieldFunction(size_t g, size_t m) const;
+
+  /**
+   * Returns the local handle to the power generation field function, if enabled.
+   */
+  size_t GetHandleToPowerGenFieldFunc() const;
+
+  void Initialize() override;
+
+  /**
+   * Initializes default materials and physics materials.
+   */
+  void InitMaterials();
+
+  /**
+   * Intializes all point sources.
+   */
+  void InitializePointSources();
+
+  void InitWGDSA(LBSGroupset& groupset, bool vaccum_bcs_are_dirichlet = true);
+  std::vector<double> WGSCopyOnlyPhi0(const LBSGroupset& groupset,
+                                      const std::vector<double>& phi_in);
+  void GSProjectBackPhi0(const LBSGroupset& groupset,
+                         const std::vector<double>& input,
+                         std::vector<double>& output);
+
+  void AssembleWGDSADeltaPhiVector(const LBSGroupset& groupset,
+                                   const std::vector<double>& phi_in,
+                                   std::vector<double>& delta_phi_local);
+
+  /**
+   * DAssembles a delta-phi vector on the first moment.
+   */
+  void DisAssembleWGDSADeltaPhiVector(const LBSGroupset& groupset,
+                                      const std::vector<double>& delta_phi_local,
+                                      std::vector<double>& ref_phi_new);
+
+  /**
+   * Assembles a delta-phi vector on the first moment.
+   */
+  void AssembleTGDSADeltaPhiVector(const LBSGroupset& groupset,
+                                   const std::vector<double>& phi_in,
+                                   std::vector<double>& delta_phi_local);
+  /**
+   * DAssembles a delta-phi vector on the first moment.
+   */
+  void DisAssembleTGDSADeltaPhiVector(const LBSGroupset& groupset,
+                                      const std::vector<double>& delta_phi_local,
+                                      std::vector<double>& ref_phi_new);
+
+  /**
+   * Writes phi_old to restart file.
+   */
+  void WriteRestartData(const std::string& folder_name, const std::string& file_base);
+
+  /**
+   * Read phi_old from restart file.
+   */
+  void ReadRestartData(const std::string& folder_name, const std::string& file_base);
+
+  /**
+   * Writes the groupset's angular fluxes to file.
+   */
+  void WriteGroupsetAngularFluxes(const LBSGroupset& groupset, const std::string& file_base);
+
+  /**
+   * Prints the groupset's angular fluxes to file.
+   */
+  void ReadGroupsetAngularFluxes(LBSGroupset& groupset, const std::string& file_base);
+
+  /**
+   * Makes a source-moments vector from scattering and fission based on the latest phi-solution.
+   */
+  std::vector<double> MakeSourceMomentsFromPhi();
+
+  /**
+   * Writes a given flux-moments vector to file.
+   */
+  void WriteFluxMoments(const std::string& file_base, const std::vector<double>& flux_moments);
+
+  /**
+   * Reads a flux-moments vector from a file in the specified vector.
+   */
+  void ReadFluxMoments(const std::string& file_base,
+                       std::vector<double>& flux_moments,
+                       bool single_file = false);
+
+  /**
+   * Copy relevant section of phi_old to the field functions.
+   */
+  void UpdateFieldFunctions();
+
+  /**
+   * Sets the internal phi vector to the value in the associated field function.
+   */
+  void SetPhiFromFieldFunctions(PhiSTLOption which_phi,
+                                const std::vector<size_t>& m_indices,
+                                const std::vector<size_t>& g_indices);
+
+  /**
+   * Compute the total fission production in the problem.
+   * \author Zachary Hardy.
+   */
+  double ComputeFissionProduction(const std::vector<double>& phi);
+
+  /**
+   * Computes the total fission rate in the problem.
+   * \author Zachary Hardy.
+   */
+  double ComputeFissionRate(const std::vector<double>& phi);
+
+  /**
+   * Compute the steady state delayed neutron precursor concentrations.
+   */
+  void ComputePrecursors();
+
+  /**
+   * Sets a value to the zeroth (scalar) moment of the vector.
+   */
+  virtual void SetPhiVectorScalarValues(std::vector<double>& phi_vector, double value);
+  /**
+   * Scales a flux moment vector. For sweep methods the delayed angular fluxes will also be scaled.
+   */
+  virtual void ScalePhiVector(PhiSTLOption which_phi, double value);
+  /**
+   * Assembles a vector for a given groupset from a source vector.
+   */
+  virtual void
+  SetGSPETScVecFromPrimarySTLvector(LBSGroupset& groupset, Vec x, PhiSTLOption which_phi);
+
+  /**
+   * Assembles a vector for a given groupset from a source vector.
+   */
+  virtual void
+  SetPrimarySTLvectorFromGSPETScVec(LBSGroupset& groupset, Vec x_src, PhiSTLOption which_phi);
+
+  /**
+   * Assembles a vector for a given groupset from a source vector.
+   */
+  virtual void GSScopedCopyPrimarySTLvectors(LBSGroupset& groupset,
+                                             const std::vector<double>& x_src,
+                                             std::vector<double>& y);
+
+  /**
+   * Assembles a vector for a given groupset from a source vector.
+   */
+  virtual void GSScopedCopyPrimarySTLvectors(LBSGroupset& groupset,
+                                             PhiSTLOption from_which_phi,
+                                             PhiSTLOption to_which_phi);
+
+  /**
+   * Assembles a vector for a given group span from a source vector.
+   */
+  virtual void SetGroupScopedPETScVecFromPrimarySTLvector(int first_group_id,
+                                                          int last_group_id,
+                                                          Vec x,
+                                                          const std::vector<double>& y);
+
+  /**
+   * Assembles a vector for a given groupset from a source vector.
+   */
+  virtual void SetPrimarySTLvectorFromGroupScopedPETScVec(int first_group_id,
+                                                          int last_group_id,
+                                                          Vec x_src,
+                                                          std::vector<double>& y);
+
+  /**
+   * Assembles a PETSc vector from multiple groupsets.
+   */
+  virtual void SetMultiGSPETScVecFromPrimarySTLvector(const std::vector<int>& gs_ids,
+                                                      Vec x,
+                                                      PhiSTLOption which_phi);
+
+  /**
+   * Disassembles a multiple Groupset PETSc vector STL vectors.
+   */
+  virtual void SetPrimarySTLvectorFromMultiGSPETScVecFrom(const std::vector<int>& gs_ids,
+                                                          Vec x_src,
+                                                          PhiSTLOption which_phi);
+
 protected:
+  /**
+   * Performs general input checks before initialization continues.
+   */
+  virtual void PerformInputChecks();
+
+  /**
+   * Prints header information of simulation.
+   */
+  void PrintSimHeader();
+
+  virtual void InitializeSpatialDiscretization();
+  void ComputeUnitIntegrals();
+
+  /**
+   * Initializes common groupset items.
+   */
+  void InitializeGroupsets();
+
+  /**
+   * Computes the number of moments for the given mesher types
+   */
+  void ComputeNumberOfMoments();
+
+  /**
+   * Initializes parallel arrays.
+   */
+  virtual void InitializeParrays();
+
+  void InitializeFieldFunctions();
+
+  void InitializeBoundaries();
+  virtual void InitializeSolverSchemes();
+  virtual void InitializeWGSSolvers(){};
+  void InitTGDSA(LBSGroupset& groupset);
+
   typedef chi_mesh::sweep_management::CellFaceNodalMapping CellFaceNodalMapping;
 
   size_t source_event_tag_ = 0;
@@ -107,232 +552,8 @@ protected:
   /**Time integration parameter meant to be set by an executor*/
   std::shared_ptr<const chi_math::TimeIntegration> time_integration_ = nullptr;
 
-public:
-  static chi::InputParameters GetInputParameters();
-  explicit LBSSolver(const std::string& text_name);
-  explicit LBSSolver(const chi::InputParameters& params);
-
-  LBSSolver(const LBSSolver&) = delete;
-  LBSSolver& operator=(const LBSSolver&) = delete;
-
-  virtual ~LBSSolver() = default;
-
-  size_t GetSourceEventTag() const;
-
-  double LastRestartWrite() const;
-  double& LastRestartWrite();
-
-  lbs::Options& Options();
-  const lbs::Options& Options() const;
-
-  static chi::InputParameters OptionsBlock();
-  static chi::InputParameters BoundaryOptionsBlock();
-  void SetOptions(const chi::InputParameters& params);
-  void SetBoundaryOptions(const chi::InputParameters& params);
-
-  size_t NumMoments() const;
-  size_t NumGroups() const;
-  size_t NumPrecursors() const;
-
-  size_t GetMaxPrecursorsPerMaterial() const;
-
-  void AddGroup(int id);
-  const std::vector<LBSGroup>& Groups() const;
-
-  void AddGroupset();
-  std::vector<LBSGroupset>& Groupsets();
-  const std::vector<LBSGroupset>& Groupsets() const;
-
-  void AddPointSource(PointSource psrc);
-  void ClearPointSources();
-  const std::vector<PointSource>& PointSources() const;
-
-  const std::map<int, XSPtr>& GetMatID2XSMap() const;
-  const std::map<int, IsotropicSrcPtr>& GetMatID2IsoSrcMap() const;
-
-  const chi_math::SpatialDiscretization& SpatialDiscretization() const;
-  const std::vector<UnitCellMatrices>& GetUnitCellMatrices() const;
-  const chi_mesh::MeshContinuum& Grid() const;
-
-  const std::vector<lbs::CellLBSView>& GetCellTransportViews() const;
-
-  std::map<uint64_t, BoundaryPreference>& BoundaryPreferences();
-
-  const chi_math::UnknownManager& UnknownManager() const;
-
-  size_t LocalNodeCount() const;
-  size_t GlobalNodeCount() const;
-
-  std::vector<double>& QMomentsLocal();
-  const std::vector<double>& QMomentsLocal() const;
-  std::vector<double>& ExtSrcMomentsLocal();
-  const std::vector<double>& ExtSrcMomentsLocal() const;
-  std::vector<double>& PhiOldLocal();
-  const std::vector<double>& PhiOldLocal() const;
-  std::vector<double>& PhiNewLocal();
-  const std::vector<double>& PhiNewLocal() const;
-  std::vector<double>& PrecursorsNewLocal();
-  const std::vector<double>& PrecursorsNewLocal() const;
-  std::vector<VecDbl>& PsiNewLocal();
-  const std::vector<VecDbl>& PsiNewLocal() const;
-
-  /**Returns the sweep boundaries as a read only reference*/
-  const std::map<uint64_t, std::shared_ptr<SweepBndry>>& SweepBoundaries() const;
-
-  SetSourceFunction GetActiveSetSourceFunction() const;
-
-  AGSLinSolverPtr GetPrimaryAGSSolver();
-
-  std::vector<LinSolvePtr>& GetWGSSolvers();
-
-  WGSContext<Mat, Vec, KSP>& GetWGSContext(int groupset_id);
-
-  virtual std::pair<size_t, size_t> GetNumPhiIterativeUnknowns();
-
-  /**Gets the local handle of a flux-moment based field function.*/
-  size_t MapPhiFieldFunction(size_t g, size_t m) const;
-
-  /**Returns the local handle to the power generation field function, if
-   * enabled.*/
-  size_t GetHandleToPowerGenFieldFunc() const;
-
-  // 01
-  void Initialize() override;
-
-protected:
-  // 01a
-  virtual void PerformInputChecks();
-  // 01b
-  void PrintSimHeader();
-
-public:
-  // 01c
-  void InitMaterials();
-
-protected:
-  // 01d
-  virtual void InitializeSpatialDiscretization();
-  void ComputeUnitIntegrals();
-  // 01e
-  void InitializeGroupsets();
-  // 01f
-  void ComputeNumberOfMoments();
-  // 01g
-  virtual void InitializeParrays();
-  //   a
-  void InitializeFieldFunctions();
-  // 01h
-  void InitializeBoundaries();
-
-public:
-  // 01i
-  void InitializePointSources();
-
-protected:
-  // 01j
-  virtual void InitializeSolverSchemes();
-  virtual void InitializeWGSSolvers(){};
-
-  // 03d
-public:
-  void InitWGDSA(LBSGroupset& groupset, bool vaccum_bcs_are_dirichlet = true);
-  std::vector<double> WGSCopyOnlyPhi0(const LBSGroupset& groupset,
-                                      const std::vector<double>& phi_in);
-  void GSProjectBackPhi0(const LBSGroupset& groupset,
-                         const std::vector<double>& input,
-                         std::vector<double>& output);
-  //  std::vector<double> WGDSAMake
-public:
-  void AssembleWGDSADeltaPhiVector(const LBSGroupset& groupset,
-                                   const std::vector<double>& phi_in,
-                                   std::vector<double>& delta_phi_local);
-
-  void DisAssembleWGDSADeltaPhiVector(const LBSGroupset& groupset,
-                                      const std::vector<double>& delta_phi_local,
-                                      std::vector<double>& ref_phi_new);
-
-protected:
   static void CleanUpWGDSA(LBSGroupset& groupset);
-
-  // 03e
-  void InitTGDSA(LBSGroupset& groupset);
-
-public:
-  void AssembleTGDSADeltaPhiVector(const LBSGroupset& groupset,
-                                   const std::vector<double>& phi_in,
-                                   std::vector<double>& delta_phi_local);
-  void DisAssembleTGDSADeltaPhiVector(const LBSGroupset& groupset,
-                                      const std::vector<double>& delta_phi_local,
-                                      std::vector<double>& ref_phi_new);
-
-protected:
   static void CleanUpTGDSA(LBSGroupset& groupset);
-  // 04 File IO
-public:
-  // 04a
-  void WriteRestartData(const std::string& folder_name, const std::string& file_base);
-  void ReadRestartData(const std::string& folder_name, const std::string& file_base);
-  // 04b
-  void WriteGroupsetAngularFluxes(const LBSGroupset& groupset, const std::string& file_base);
-  void ReadGroupsetAngularFluxes(LBSGroupset& groupset, const std::string& file_base);
-
-  // 04c
-  std::vector<double> MakeSourceMomentsFromPhi();
-  void WriteFluxMoments(const std::string& file_base, const std::vector<double>& flux_moments);
-  void ReadFluxMoments(const std::string& file_base,
-                       std::vector<double>& flux_moments,
-                       bool single_file = false);
-
-  // 05a
-  void UpdateFieldFunctions();
-  void SetPhiFromFieldFunctions(PhiSTLOption which_phi,
-                                const std::vector<size_t>& m_indices,
-                                const std::vector<size_t>& g_indices);
-
-  // 06b
-public:
-  double ComputeFissionProduction(const std::vector<double>& phi);
-  double ComputeFissionRate(const std::vector<double>& phi);
-
-  // 06c
-public:
-  void ComputePrecursors();
-
-  // 07 Vector assembly
-public:
-  virtual void SetPhiVectorScalarValues(std::vector<double>& phi_vector, double value);
-  virtual void ScalePhiVector(PhiSTLOption which_phi, double value);
-  virtual void
-  SetGSPETScVecFromPrimarySTLvector(LBSGroupset& groupset, Vec x, PhiSTLOption which_phi);
-
-  virtual void
-  SetPrimarySTLvectorFromGSPETScVec(LBSGroupset& groupset, Vec x_src, PhiSTLOption which_phi);
-
-  virtual void GSScopedCopyPrimarySTLvectors(LBSGroupset& groupset,
-                                             const std::vector<double>& x_src,
-                                             std::vector<double>& y);
-
-  virtual void GSScopedCopyPrimarySTLvectors(LBSGroupset& groupset,
-                                             PhiSTLOption from_which_phi,
-                                             PhiSTLOption to_which_phi);
-
-  virtual void SetGroupScopedPETScVecFromPrimarySTLvector(int first_group_id,
-                                                          int last_group_id,
-                                                          Vec x,
-                                                          const std::vector<double>& y);
-
-  virtual void SetPrimarySTLvectorFromGroupScopedPETScVec(int first_group_id,
-                                                          int last_group_id,
-                                                          Vec x_src,
-                                                          std::vector<double>& y);
-
-  virtual void SetMultiGSPETScVecFromPrimarySTLvector(const std::vector<int>& gs_ids,
-                                                      Vec x,
-                                                      PhiSTLOption which_phi);
-
-  virtual void SetPrimarySTLvectorFromMultiGSPETScVecFrom(const std::vector<int>& gs_ids,
-                                                          Vec x_src,
-                                                          PhiSTLOption which_phi);
 };
 
 } // namespace lbs
